+++ src/components/Layout/Layout.tsx
@@ -1,18 +1,21 @@
 import React, { useEffect, useState } from "react";
 import {
   DSProvider,
-  Header,
   SkeletonLoader,
 } from "@nypl/design-system-react-components";
 import Feedback from "~/src/components/Feedback/Feedback";
 import { useRouter } from "next/router";
+import { Header, navConfig } from "@nypl/dgx-header-component";
 
 /**
  * Container class providing header, footer,
  * and other set up information to all its children.
  */
 
-const Layout: React.FC = ({ children }) => {
+const Layout: React.FC<{ children; isTestMode?: boolean }> = ({
+  children,
+  isTestMode = false,
+}) => {
   const router = useRouter();
   const [loading, setLoading] = useState(true);
 
@@ -37,7 +40,13 @@ const Layout: React.FC = ({ children }) => {
   return (
     <>
       <DSProvider>
-        <Header />
+        {!isTestMode && (
+          <Header
+            urlType="absolute"
+            skipNav={{ target: "main-content" }}
+            navData={navConfig.current}
+          />
+        )}
         <main id="main-content">
           {router.isFallback || loading ? (
             <>