+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGE LOG
 
+## [Pre-release]
+- Add github actions to automate tag and release
+- Replace DS Header component with dgx-header-component
+
 ## [0.15.1]
 
 - Fix absolute link is prefixed by host issue